\name{calibQuantiles-package}
\alias{calibQuantiles-package}
\alias{calibQuantiles}
\docType{package}
\title{
\packageTitle{calibQuantiles}
}
\description{
\packageDescription{calibQuantiles}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{calibQuantiles}
\packageIndices{calibQuantiles}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{calibQuantiles}

Maintainer: \packageMaintainer{calibQuantiles}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
